Fix updates of the entities by saving IVs.
Server checks that values of final fields didn't change. In case of
encrypted fields, we've been re-encrypting them with new random IV. To
get the same value, we must store and re-use the IVs. Current design
adds a map to each entity with saved IVs which serializer uses. This is
suboptimal because in most cases we don't want to read and parse them.

Slightly more effective solution would be to put them not on all
aggregates but just in one map on the top level of the entity with keys
being object ids and values being IVs. We wouldn't read them in most
cases then.

Another approach would be to store them in a separate table with one
row per top-level entity. Then we would explicitly request them for
updates only but we would need to write them in transaction.

Doing these optimizations would complicate serialization because we
would need to look up fields by the entity id.