[android] Use Dagger for user-scoped dependency injection
Bye-bye DependencyDump!

Current refactoring lays down foundation for fearless multi-user
environment. For secure app like this we don't want users' data mixed
up.

We use scoped injection to differentiate between components of
different users. It doesn't depend on the loaded user (it might be
later) but on the selected user. So for resume when we already know
the passphrase we don't wait for full session resume to start a scope.

We might use another scope or inject fully initialized API as Deferred
to not require all clients to wait for login just in case.

It might also be a good idea to introduce coroutine scopes which
depend on the (local) user session.