[bridge] Fix LIST responses
We didn't handle `LIST "" ""` and list with `%` correctly. The first
one shouldn't match anything in our case (as there's no root) and should
return hierarchy delimiter. The second is wildcard without hierarchy
and we didn't treat it as such. All of this led to Geary not being able
to understand where's our INBOX. We also switched hierarchy delimiter
to `/` as `#` has some different special meaning by convention.